backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
ultralytics==8.3.55
Pillow==11.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
numpy==2.1.3
opencv-python-headless==4.10.0.84
